requirements.txt
torch
torchvision
pillow
tqdm
fastapi
uvicorn[standard]
jinja2
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
SQLAlchemy
bcrypt

